login-register done, message func done

+++ studycom/views.py
@@ -1,7 +1,58 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+
+from .models import Room, Topic, Message
 from .forms import RoomForm
+
+
+def login_view(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'User does not exist')
+
+
+    context = {}
+    return render(request, 'studycom/login_page.html', context)
+
+
+def register_user(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration')
+
+    context = {'form': form}
+    return render(request, 'studycom/register_page.html', context)
+
+
+def logout_view(request):
+    logout(request)
+
+    return redirect('home')
 
 
 def home_view(request):
@@ -22,12 +73,26 @@ def home_view(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
 
-    context = {'room': room}
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body'),
+        )
+        room.participants.add(request.user)
 
+        message.save()
+        return redirect('room', pk=room.id)
+
+
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
     return render(request, 'studycom/room.html', context)
 
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -40,9 +105,13 @@ def create_room(request):
     return render(request, 'studycom/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -54,10 +123,29 @@ def update_room(request, pk):
     return render(request, 'studycom/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
 
     return render(request, 'studycom/delete.html', {'obj': room})
+
+
+@login_required(login_url='login')
+def delete_message(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed to edit')
+    
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+
+    return render(request, 'studycom/delete.html', {'obj': message})
